Clipboard Mask, Tracker Icon Notifier
Mask working at
-Tracker Screen
-Pattern Screen

Tracker can give button combo feedback on tracker screen, to be extended to other screens which may benefit from that ux mechanic

+++ modules/gpu/gpu.hpp
@@ -131,8 +131,8 @@ class Gpu {
 	const unsigned short *MAP2;
 
 	VirtualScreen *vs;
-	int vcount;
-	bool blink;
+	u16 	vcount;
+	bool 	blink;
 	
 	Gpu(void);
